+++ cli/cli/Commands/InitCommand.cs
@@ -119,7 +119,7 @@ public class InitCommand : AtomicCommand<InitCommandArgs, InitCommandResult>,
 
 		{ // set the host string from existing value or given parameter, and then reset cid/pid
 			host = _configService.SetConfigString(Constants.CONFIG_PLATFORM, GetHost(args));
-			_ctx.Set(string.Empty, string.Empty, host);
+			await _ctx.Set(string.Empty, string.Empty, host);
 		}
 		
 		SaveExtraPathFiles(args);
@@ -209,11 +209,11 @@ public class InitCommand : AtomicCommand<InitCommandArgs, InitCommandResult>,
 		SaveExtraPathFiles(args);
 
 		if (!_retry) AnsiConsole.Write(new FigletText("Beam").Color(Color.Red));
-		else _ctx.Set(string.Empty, string.Empty, _ctx.Host);
+		else await _ctx.Set(string.Empty, string.Empty, _ctx.Host);
 
 		var host = _configService.SetConfigString(Constants.CONFIG_PLATFORM, GetHost(args));
 		var cid = await GetCid(args);
-		_ctx.Set(cid, string.Empty, host);
+		await _ctx.Set(cid, string.Empty, host);
 
 		if (!AliasHelper.IsCid(cid))
 		{
@@ -272,7 +272,7 @@ public class InitCommand : AtomicCommand<InitCommandArgs, InitCommandResult>,
 		//  then just returning the pid and auth
 		if (!string.IsNullOrEmpty(_ctx.Pid) && string.IsNullOrEmpty(args.pid))
 		{
-			_ctx.Set(cid, _ctx.Pid, host);
+			await _ctx.Set(cid, _ctx.Pid, host);
 			_configService.SetBeamableDirectory(_ctx.WorkingDirectory);
 			_configService.FlushConfig();
 			_configService.CreateIgnoreFile();
@@ -285,7 +285,7 @@ public class InitCommand : AtomicCommand<InitCommandArgs, InitCommandResult>,
 		// If we have a given pid, let's login there.
 		if (!string.IsNullOrEmpty(args.pid))
 		{
-			_ctx.Set(cid, args.pid, host);
+			await _ctx.Set(cid, args.pid, host);
 
 			
 			var didLogin = !args.SaveToFile || await Login(args);
@@ -304,7 +304,7 @@ public class InitCommand : AtomicCommand<InitCommandArgs, InitCommandResult>,
 			return didLogin;
 		}
 
-		_ctx.Set(cid, null, host);
+		await _ctx.Set(cid, null, host);
 		_configService.SetBeamableDirectory(_ctx.WorkingDirectory);
 		_configService.FlushConfig();
 		_configService.CreateIgnoreFile();
@@ -316,7 +316,7 @@ public class InitCommand : AtomicCommand<InitCommandArgs, InitCommandResult>,
 			return false;
 		}
 
-		_ctx.Set(cid, pid, host);
+		await _ctx.Set(cid, pid, host);
 		_configService.SetConfigString(Constants.CONFIG_PID, pid);
 		_configService.FlushConfig();