+++ cli/cli/Commands/Project/GenerateClientFileCommand.cs
@@ -43,7 +43,9 @@ public class GenerateClientFileCommand
 	: AppCommand<GenerateClientFileCommandArgs>
 		, IResultSteam<DefaultStreamResultChannel, GenerateClientFileEvent>
 {
-	public GenerateClientFileCommand() : base("generate-client", "Generate a C# client file based on a built C# microservice dll directory")
+	public override bool IsForInternalUse => true;
+
+	public GenerateClientFileCommand() : base("generate-client", "Obsolete command, please use generate-client-oapi that used the OpenAPI specifications to generate the C# client code. The generate-client command will Generate a C# client file based on a built C# microservice dll directory using refactor")
 	{
 	}