+++ cli/cli/Services/IAppContext.cs
@@ -1,5 +1,6 @@
 using Beamable.Common.Api;
 using Beamable.Common.Api.Auth;
+using Beamable.Common.Api.Realms;
 using Beamable.Common.Dependencies;
 using Beamable.Server.Common;
 using cli.Options;
@@ -31,6 +32,7 @@ public interface IAppContext : IRealmInfo
 	public bool ShowRawOutput { get; }
 	public bool ShowPrettyOutput { get; }
 	public string DotnetPath { get; }
+	public HashSet<string> IgnoreBeamoIds { get; }
 	public string WorkingDirectory { get; }
 	public IAccessToken Token { get; }
 	public string RefreshToken { get; }
@@ -61,9 +63,9 @@ public interface IAppContext : IRealmInfo
 	/// As we add more context variables, this method is responsible for "figuring them out"
 	/// </summary>
 	/// <param name="bindingContext"></param>
-	void Apply(BindingContext bindingContext);
+	Task Apply(BindingContext bindingContext);
 
-	void Set(string cid, string pid, string host);
+	Task Set(string cid, string pid, string host);
 	void UpdateToken(TokenResponse response);
 }
 
@@ -96,6 +98,7 @@ public class DefaultAppContext : IAppContext
 
 	public string DotnetPath { get; private set; }
 	public string DockerPath { get; private set; }
+	public HashSet<string> IgnoreBeamoIds { get; private set; }
 
 
 	/// <inheritdoc cref="IAppContext.ExecutingVersion"/>
@@ -151,6 +154,7 @@ public class DefaultAppContext : IAppContext
 	private string _cid, _pid, _host;
 	private string _refreshToken;
 	private BindingContext _bindingContext;
+	private readonly IAliasService _aliasService;
 	public string Cid => _cid;
 	public string Pid => _pid;
 	public string Host => _host;
@@ -163,7 +167,7 @@ public class DefaultAppContext : IAppContext
 		ConfigService configService, CliEnvironment environment, ShowRawOutput showRawOption, SkipStandaloneValidationOption skipValidationOption,
 		DotnetPathOption dotnetPathOption, ShowPrettyOutput showPrettyOption, BeamLogSwitch logSwitch,
 		UnmaskLogsOption unmaskLogsOption, NoLogFileOption noLogFileOption, DockerPathOption dockerPathOption,
-		PreferRemoteFederationOption routeMapOption)
+		PreferRemoteFederationOption routeMapOption, IAliasService aliasService)
 	{
 		_consoleContext = consoleContext;
 		_dryRunOption = dryRunOption;
@@ -184,7 +188,10 @@ public class DefaultAppContext : IAppContext
 		_routeMapOption = routeMapOption;
 		_skipValidationOption = skipValidationOption;
 		_dotnetPathOption = dotnetPathOption;
+		_aliasService = aliasService;
 		DockerPath = consoleContext.ParseResult.GetValueForOption(dockerPathOption);
+		IgnoreBeamoIds =
+			new HashSet<string>(consoleContext.ParseResult.GetValueForOption(IgnoreBeamoIdsOption.Instance));
 	}
 
 	void SetupOutputStrategy()
@@ -258,7 +265,7 @@ public class DefaultAppContext : IAppContext
 		}
 	}
 
-	public void Apply(BindingContext bindingContext)
+	public async Task Apply(BindingContext bindingContext)
 	{
 		_bindingContext = bindingContext;
 		ShowRawOutput = bindingContext.ParseResult.GetValueForOption(_showRawOption);
@@ -284,19 +291,19 @@ public class DefaultAppContext : IAppContext
 
 		_configService.Init(bindingContext);
 
-		if (!_configService.TryGetSetting(out _cid, bindingContext, _cidOption))
+		if (!_configService.TryGetSetting(out string cid, bindingContext, _cidOption))
 		{
 			// throw new CliException("cannot run without a cid. Please login.");
 		}
 
-		if (!_configService.TryGetSetting(out _pid, bindingContext, _pidOption))
+		if (!_configService.TryGetSetting(out string pid, bindingContext, _pidOption))
 		{
 			// throw new CliException("cannot run without a cid. Please login.");
 		}
 
-		if (!_configService.TryGetSetting(out _host, bindingContext, _hostOption))
+		if (!_configService.TryGetSetting(out string host, bindingContext, _hostOption))
 		{
-			_host = Constants.DEFAULT_PLATFORM;
+			host = Constants.DEFAULT_PLATFORM;
 			// throw new CliException("cannot run without a cid. Please login.");
 		}
 
@@ -314,15 +321,24 @@ public class DefaultAppContext : IAppContext
 		_configService.TryGetSetting(out var accessToken, bindingContext, _accessTokenOption, defaultAccessToken);
 		_configService.TryGetSetting(out _refreshToken, bindingContext, _refreshTokenOption, defaultRefreshToken);
 
-		_token = new CliToken(accessToken, RefreshToken, _cid, _pid);
-		Set(_cid, _pid, _host);
+		_token = new CliToken(accessToken, RefreshToken, cid, pid);
+		await Set(cid, pid, host);
 		
 
 	}
 
-	public void Set(string cid, string pid, string host)
+	public async Task Set(string cid, string pid, string host)
 	{
-		_cid = cid;
+		if (!string.IsNullOrEmpty(cid))
+		{
+			var aliasResolve = await _aliasService.Resolve(cid);
+			_cid = aliasResolve.Cid;
+		}
+		else
+		{
+			_cid = cid;
+		}
+
 		_pid = pid;
 		_host = host;
 		_token.Cid = _cid;