+++ cli/cli/Services/UnrealSourceGenerator/Declarations/UnrealEnumDeclaration.cs
@@ -6,6 +6,7 @@ public struct UnrealEnumDeclaration
 {
 	public UnrealType UnrealTypeName;
 	public NamespacedType NamespacedTypeName;
+	public string ServiceName;
 
 	public List<string> EnumValues;
 
@@ -22,6 +23,7 @@ public struct UnrealEnumDeclaration
 		helperDict.Add(nameof(UnrealTypeName), UnrealTypeName);
 		helperDict.Add(nameof(NamespacedTypeName), NamespacedTypeName);
 		helperDict.Add(nameof(EnumValues), enumValues);
+		helperDict.Add(nameof(ServiceName), ServiceName);
 	}
 
 	public const string U_ENUM_HEADER = $@"#pragma once
@@ -30,19 +32,19 @@ public struct UnrealEnumDeclaration
 
 #include ""₢{nameof(NamespacedTypeName)}₢.generated.h""
 
-UENUM(BlueprintType, Category=""Beam|Enums"")
+UENUM(BlueprintType, Category=""Beam|₢{nameof(ServiceName)}₢|Utils|Enums"")
 enum class ₢{nameof(UnrealTypeName)}₢ : uint8
 {{
 	₢{nameof(EnumValues)}₢		
 }};
 
-UCLASS(BlueprintType, Category=""Beam|Enums"")
+UCLASS(BlueprintType, Category=""Beam|₢{nameof(ServiceName)}₢|Utils|Enums"")
 class ₢{nameof(exportMacro)}₢ U₢{nameof(NamespacedTypeName)}₢Library : public UBlueprintFunctionLibrary
 {{
 	GENERATED_BODY()
 public:		
 	
-	UFUNCTION(BlueprintPure, meta = (DisplayName = ""Beam - ₢{nameof(NamespacedTypeName)}₢ To Serialization Name"", CompactNodeTitle = ""->""), Category=""Beam|Enums"")
+	UFUNCTION(BlueprintPure, meta = (DisplayName=""₢{nameof(NamespacedTypeName)}₢ To Serialization Name"", CompactNodeTitle = ""->""), Category=""Beam|₢{nameof(ServiceName)}₢|Utils|Enums"")
 	static FString ₢{nameof(NamespacedTypeName)}₢ToSerializationName(₢{nameof(UnrealTypeName)}₢ Value)
 	{{
 		const UEnum* Enum = StaticEnum<₢{nameof(UnrealTypeName)}₢>();
@@ -54,7 +56,7 @@ public:
 		
 	}}
 
-	UFUNCTION(BlueprintPure, meta = (DisplayName = ""Beam - Serialization Name To ₢{nameof(NamespacedTypeName)}₢"", CompactNodeTitle = ""->""), Category=""Beam|Enums"")
+	UFUNCTION(BlueprintPure, meta = (DisplayName=""Serialization Name To ₢{nameof(NamespacedTypeName)}₢"", CompactNodeTitle = ""->""), Category=""Beam|₢{nameof(ServiceName)}₢|Utils|Enums"")
 	static ₢{nameof(UnrealTypeName)}₢ SerializationNameTo₢{nameof(NamespacedTypeName)}₢(FString Value)
 	{{
 		const UEnum* Enum = StaticEnum<₢{nameof(UnrealTypeName)}₢>();