+++ client/Packages/com.beamable.server/SharedRuntime/Api/Analytics/IMicroserviceAnalyticsService.cs
@@ -1,0 +1,11 @@
+// this file was copied from nuget package Beamable.Server.Common@4.3.0-PREVIEW.RC2
+// https://www.nuget.org/packages/Beamable.Server.Common/4.3.0-PREVIEW.RC2
+
+﻿using Beamable.Api.Analytics;
+
+namespace Beamable.Server.Api.Analytics
+{
+	public interface IMicroserviceAnalyticsService : IBeamAnalyticsService
+	{
+	}
+}