+++ client/Packages/com.beamable.server/SharedRuntime/Api/Payments/IMicroservicePaymentsApi.cs
@@ -1,0 +1,12 @@
+// this file was copied from nuget package Beamable.Server.Common@4.3.0-PREVIEW.RC2
+// https://www.nuget.org/packages/Beamable.Server.Common/4.3.0-PREVIEW.RC2
+
+﻿using Beamable.Common.Api.Payments;
+
+namespace Beamable.Server.Api.Payments
+{
+	public interface IMicroservicePaymentsApi : IPaymentsApi
+	{
+
+	}
+}