+++ client/Packages/com.beamable.server/SharedRuntime/StorageObject.cs
@@ -1,0 +1,15 @@
+// this file was copied from nuget package Beamable.Server.Common@4.3.0-PREVIEW.RC2
+// https://www.nuget.org/packages/Beamable.Server.Common/4.3.0-PREVIEW.RC2
+
+namespace Beamable.Server
+{
+	public abstract class StorageObject
+	{
+
+	}
+
+	public abstract class MongoStorageObject : StorageObject
+	{
+
+	}
+}