+++ client/Packages/com.beamable/Common/Runtime/BeamCli/ServiceInfoHelper.cs
@@ -1,5 +1,5 @@
-// this file was copied from nuget package Beamable.Common@4.2.0
-// https://www.nuget.org/packages/Beamable.Common/4.2.0
+// this file was copied from nuget package Beamable.Common@4.3.0-PREVIEW.RC2
+// https://www.nuget.org/packages/Beamable.Common/4.3.0-PREVIEW.RC2
 
 ﻿using Beamable.Common.BeamCli.Contracts;
 using System.IO;