+++ client/Packages/com.beamable/CHANGELOG.md
@@ -5,17 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-##[Unreleased]
+## [Unreleased] 
+
+### Fixed
+- Fixed issue that CloudSaving could generate corrupted save files if the application was closed during saving process.
+### Change
+- Able to use the new Client Code Generator from CLI that uses OpenAPI instead of the old one that uses Reflection
+
+
+## [2.3.0] - 2025-05-08
 
 ### Added
 - Helper menu option for updating Realm configuration for GPGS integration.
+- Information about already installed versions in information about missing Dotnet.
+- `Open portal for user` button in `BeamableBehaviour` inspector.
 
 ### Changed
+- Upgrade CLI to 4.3.0
 - `AccountManagementConfiguration` no longer overrides `BEAMABLE_GPGS` define symbol.
+- Improved IAP Purchaser initialization flow.
 
 ### Fixed
 - CID/PID Mismatch error message was too big for Unity popup. Now it uses a Beamable Custom Editor Window. [3933](https://github.com/beamable/BeamableProduct/issues/3933)
 - Improved error detection for attaching identity providers in `PlayerAccounts`.
+- Improved content selection in Content window
 
 ## [2.2.0] - 2025-04-04
 ### Added